[REFACTOR] 미사용 클래스 제거 패키지 구조 변경
