+++ src/main/java/com/luckytree/shop_service/shop/application/service/CreateShopService.java
@@ -1,11 +1,8 @@
 package com.luckytree.shop_service.shop.application.service;
 
 import com.luckytree.shop_service.shop.application.port.in.CreateShopUseCase;
-import com.luckytree.shop_service.shop.domain.ShopRequest;
+import com.luckytree.shop_service.shop.application.port.in.ShopRequest;
 import com.luckytree.shop_service.shop.application.port.out.CreateShopPort;
-import com.luckytree.shop_service.shop.domain.ShopDetail;
-import com.luckytree.shop_service.shop.domain.ShopSummary;
-import com.luckytree.shop_service.shop.domain.ShopStatus;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
