+++ src/main/java/com/luckytree/shop_service/shop/domain/ShopSummary.java
@@ -16,25 +16,6 @@ public class ShopSummary extends BaseTimeEntity {
     private String category;
     private Hashtag hashtag;
 
-    @Builder
-    public ShopSummary(Long id, String name, ShopStatus status, Double mapX, Double mapY, String category, Hashtag hashtag) {
-        this.id = id;
-        this.name = name;
-        this.status = status;
-        this.mapX = mapX;
-        this.mapY = mapY;
-        this.category = category;
-        this.hashtag = hashtag;
-    }
-
-    public ShopSummary(ShopRequest shopRequest, ShopStatus status) {
-        this.name = shopRequest.getShopName();
-        this.status = status;
-        this.mapX = shopRequest.getMapX();
-        this.mapY = shopRequest.getMapY();
-        this.category = shopRequest.getCategory();
-    }
-
     public ShopSummary(ShopEntity shopEntity) {
         this.id = shopEntity.getId();
         this.name = shopEntity.getName();
